+++ brest/cameras/__init__.py
@@ -8,4 +8,4 @@ __all__ = [
     'PointGrey'
     ]
 
-__version__ = '0.0.8'
+__version__ = '0.0.9'