+++ brest/communication/__init__.py
@@ -37,4 +37,4 @@ __all__ = [
     'types',
     ]
 
-__version__ = '0.0.8'
+__version__ = '0.0.9'