+++ brest/flashers/__init__.py
@@ -4,4 +4,4 @@ from .jlink import JLink
 
 __all__ = ['Flashers', 'STLink', 'JLink']
 
-__version__ = '0.0.8'
+__version__ = '0.0.9'