+++ brest/interfaces/__init__.py
@@ -6,4 +6,4 @@ __all__ = [
     'SerialInterface',
 ]
 
-__version__ = '0.0.8'
+__version__ = '0.0.9'